[microNPU] Introduce a pass to remove redundant identity operations
Introduces a  pass that aims to remove identity operations,
introduced during legalization, that are immediately followed by an NPU
compute operation e.g. Convolution.
