+++ tests/python/contrib/test_ethosu/test_identity_optimizer.py
@@ -1,0 +1,245 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+"""
+Test the identity optimizer pass that removes redundant identity
+operations from the microNPU codegen.
+"""
+import numpy as np
+import tensorflow as tf
+
+import tvm
+from tvm import relay
+from tvm.relay.op.contrib.ethosu import partition_for_ethosu
+from tvm.relay.backend.contrib.ethosu.codegen import relay_to_tir_func
+from tvm.relay.backend.contrib.ethosu.codegen import IdentityOptimizer
+
+from . import infra
+from .test_codegen import _compare_tvm_with_tflite
+
+
+def _optimize(expr, optimize=True):
+    """Create IRModule and run layout optimizer pass."""
+    mod = tvm.IRModule.from_expr(expr)
+    mod = relay.transform.InferType()(mod)
+    if optimize:
+        mod = IdentityOptimizer()(mod)
+    print(mod)
+    entry = mod["main"]
+    return entry if isinstance(expr, relay.Function) else entry.body
+
+
+def _assert_structural_equal(a, b):
+    """Check structural equality of two Relay expressions."""
+    reason = (
+        "Actual and expected relay functions are not equal. "
+        "IdentityOptimizer is not correctly removing redundant "
+        "identity operations."
+    )
+    assert tvm.ir.structural_equal(a, b), reason
+
+
+def test_simple_reshape_identity_removal():
+    """Check identity is removed when there is a reshape in
+    the graph and a compute operation follows."""
+
+    def get_graph(get_expected=False):
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        x = infra.make_ethosu_conv2d(x, 4, 4, (1, 1), (0, 0), (1, 1), (1, 1))
+        x = relay.reshape(x, newshape=(1, 4, 4, 1))
+        if not get_expected:
+            x = infra.make_ethosu_identity(x)
+        x = infra.make_ethosu_unary_elementwise(x, 1, "ABS")
+        return relay.Function(relay.analysis.free_vars(x), x)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(get_expected=True), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_simple_strided_slice_identity_removal():
+    """Check identity is removed when there is a strided slice
+    in the graph and a compute operation follows."""
+
+    def get_graph(get_expected=False):
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        x = infra.make_ethosu_pooling(x, "MAX", (1, 1), 4, (1, 1), (0, 0))
+        x = relay.strided_slice(x, begin=[0, 0, 0, 0], end=[1, 2, 2, 2])
+        if not get_expected:
+            x = infra.make_ethosu_identity(x)
+        x = infra.make_ethosu_pooling(x, "MAX", (1, 1), 2, (1, 1), (0, 0))
+        return relay.Function(relay.analysis.free_vars(x), x)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(get_expected=True), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_no_identity():
+    """Check the graph is not affected when there is no identity in the graph."""
+
+    def get_graph():
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        x = infra.make_ethosu_conv2d(x, 4, 4, (1, 1), (0, 0), (1, 1), (1, 1))
+        x = infra.make_ethosu_pooling(x, "MAX", (1, 1), 4, (1, 1), (0, 0))
+        x = infra.make_ethosu_depthwise_conv2d(x, 4, (1, 1), (0, 0), (1, 1), (1, 1))
+        x = infra.make_ethosu_unary_elementwise(x, 4, "ABS")
+        return relay.Function(relay.analysis.free_vars(x), x)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_reshape_last():
+    """Check that an identity as a leaf of the graph is not removed."""
+
+    def get_graph():
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        x = infra.make_ethosu_conv2d(x, 4, 4, (1, 1), (0, 0), (1, 1), (1, 1))
+        x = relay.reshape(x, newshape=(1, 4, 4, 1))
+        x = infra.make_ethosu_identity(x)
+        return relay.Function(relay.analysis.free_vars(x), x)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_requantize_identity_no_removal():
+    """Check that an identity that actually performs a requantize isn't removed."""
+
+    def get_graph():
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        x = relay.reshape(x, newshape=(1, 1, 4, 4))
+        x = infra.make_ethosu_identity(
+            x, ifm_scale=0.5, ifm_zero_point=1, ofm_scale=0.3, ofm_zero_point=2
+        )
+        x = infra.make_ethosu_unary_elementwise(x, 4, "ABS")
+        return relay.Function(relay.analysis.free_vars(x), x)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_multiple_output_identity():
+    """Check that an identity isn't removed when it has multiple outputs,
+    as this is not supported yet."""
+
+    def get_graph():
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        x = infra.make_ethosu_identity(x)
+        y = infra.make_ethosu_unary_elementwise(x, 4, "ABS")
+        z = infra.make_ethosu_unary_elementwise(x, 4, "ABS")
+        out = relay.concatenate((y, z), axis=0)
+        return relay.Function(relay.analysis.free_vars(x), out)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_identity_removal_with_multiple_transform_ops():
+    """Check that only an identity directly parent to a compute
+    operation is removed."""
+
+    def get_graph(get_expected=False):
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        x = relay.strided_slice(x, begin=[0, 0, 0, 0], end=[1, 2, 2, 2])
+        x = infra.make_ethosu_identity(x)
+        x = relay.reshape(x, newshape=(1, 1, 1, 8))
+        if not get_expected:
+            x = infra.make_ethosu_identity(x)
+        x = infra.make_ethosu_unary_elementwise(x, 8, "ABS")
+        return relay.Function(relay.analysis.free_vars(x), x)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(get_expected=True), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_identity_removal_on_binary_elementwise():
+    """Check identities before binary elementwise are removed correctly."""
+
+    def get_graph(get_expected=False):
+        x = relay.var("x", shape=(1, 2, 2, 4), dtype="int8")
+        y = relay.var("y", shape=(1, 2, 2, 4), dtype="int8")
+        if not get_expected:
+            x = infra.make_ethosu_identity(x)
+            y = infra.make_ethosu_identity(y)
+        z = infra.make_ethosu_binary_elementwise(x, y, 4, 4, "ADD", "int8")
+        return relay.Function(relay.analysis.free_vars(z), z)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(get_expected=True), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_identity_single_removal_on_binary_elementwise():
+    """Check that identity on the second input of the binary elementwise
+    operation is removed while the other input has no identity."""
+
+    def get_graph(get_expected=False):
+        x = relay.var("x", shape=(1, 4, 1, 4), dtype="int8")
+        y = relay.var("y", shape=(1, 2, 2, 4), dtype="int8")
+        y = relay.reshape(y, newshape=(1, 4, 1, 4))
+        if not get_expected:
+            y = infra.make_ethosu_identity(y)
+        z = infra.make_ethosu_binary_elementwise(x, y, 4, 4, "ADD", "int8")
+        return relay.Function(relay.analysis.free_vars(z), z)
+
+    actual = _optimize(get_graph())
+    expected = _optimize(get_graph(get_expected=True), optimize=False)
+    _assert_structural_equal(actual, expected)
+
+
+def test_layout_optimizer_runs_in_compilation_pipeline():
+    """Checks that the identity optimization pass is run as part of the NPU compilation pipeline."""
+
+    def get_graph():
+        x = relay.var("x", shape=(1, 4, 4, 4), dtype="int8")
+        x = relay.reshape(x, newshape=(1, 1, 16, 4))
+        x = relay.nn.max_pool2d(x, layout="NHWC")
+        func = relay.Function(relay.analysis.free_vars(x), x)
+        return tvm.IRModule.from_expr(func)
+
+    mod = get_graph()
+    mod = partition_for_ethosu(mod)
+
+    external_gv_name = mod["main"].body.op.name_hint
+    external_func = mod[external_gv_name]
+    prim_func = relay_to_tir_func(external_func)
+
+    # Check for hints in the TIR prim func that the identity optimization pass
+    # has ran. There should not be an identity in the prim func.
+    assert prim_func.body.value.args[0] == "ethosu_pooling"
+
+
+def test_same_output():
+    """Check that the output remains the same when the identity
+    optimizer pass removes some identities inserted during legalization."""
+    ifm_shapes = [(1, 1, 25, 8), (1, 25, 1, 8)]
+
+    @tf.function
+    def model(x, y):
+        x = tf.reshape(x, (1, 5, 5, 8))
+        z = tf.add(x, y)
+        z = tf.reshape(z, (1, 1, 25, 8))
+        return z
+
+    _compare_tvm_with_tflite(model, ifm_shapes, "ethos-u55-256")
